Feature: Run Uboot in cache
Flow: SPL (EL3), Uboot (EL3), ATF (EL3), Uboot (EL2)

What happens when:
SPL loads Uboot to cache then branches over, Uboot runs in
cache, initializes the main memory and flushes the caches.
Then it branches over to ATF that returns to Uboot switching
the Exception Level from 3 to 2 (EL3 to EL2). Then the system
boots as usual.

In this flavour ATF is linked in to Uboot through the linker script.

Note: this Uboot requires amended ATF that can be found in the ATF
git repo on the same branch name.

To build all in one (SPL, Uboot with ATF in) run:
./scripts/scripts/build_uboot_run_in_cache.sh

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>