board/axxia: Updates to NetBoot on 6700
Support gmac48-51, provide a macstats tool, clear the memory
EIOA uses before and not after the EIOA takes away, print
the memory EIOA uses to the console (if enabled through
a config) and to bdinfo, add an option to allocate an enormous
EIOA struct (over 1M) either to the heap or .bss section
(by default the struct is in .bss), fix catching an Abort
signal (Ctrl+C), eg. during tftp download.

Signed-off-by: Marek Bykowski <marekx.bykowski@intel.com>